+++ package.json
@@ -22,6 +22,8 @@
   "devDependencies": {
     "@crxjs/vite-plugin": "^1.0.14",
     "@preact/preset-vite": "^2.1.5",
+    "@types/chrome": "^0.0.263",
+    "typescript": "^5.4.3",
     "vite": "^2.9.9"
   }
 }
