+++ src/Utils/utils.ts
@@ -1,3 +1,3 @@
-export function sleep(ms) {
+export function sleep(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
